move RTSP-related classes into module
