+++ rosys/vision/__init__.py
@@ -9,8 +9,7 @@ from .detector_hardware import DetectorHardware
 from .detector_simulation import DetectorSimulation, SimulatedObject
 from .image import Image, ImageSize
 from .multi_camera_provider import MultiCameraProvider
-from .rtsp_camera import RtspCamera
-from .rtsp_camera_provider import RtspCameraProvider
+from .rtsp_camera import RtspCamera, RtspCameraProvider
 from .simulated_camera import SimulatedCamera
 from .simulated_camera_provider import SimulatedCameraProvider, SimulatedDeviceScanner
 from .usb_camera import UsbCamera
