+++ rosys/vision/rtsp_camera/__init__.py
@@ -1,0 +1,7 @@
+from .rtsp_camera import RtspCamera
+from .rtsp_camera_provider import RtspCameraProvider
+
+__all__ = [
+    'RtspCamera',
+    'RtspCameraProvider',
+]
