+++ rosys/vision/rtsp_camera/rtsp_camera.py
@@ -1,40 +1,21 @@
-import asyncio
 import io
 import logging
 import os
-import shlex
-import subprocess
 import sys
-from dataclasses import dataclass, field
-from io import BytesIO
-from typing import Any, AsyncGenerator, Coroutine, Optional
+from typing import Any, Optional
 
-import cv2
 import imgsize
 import netifaces
 import PIL
 
-from .. import persistence, rosys
-from ..geometry import Rectangle
-from .camera import Camera, ConfigurableCameraMixin, TransformableCameraMixin
-from .image import Image, ImageSize
-from .image_processing import process_jpeg_image
-from .image_rotation import ImageRotation
-from .jovision_rtsp_interface import JovisionInterface
-
-
-class VendorType:
-    JOVISION = 1
-    DAHUA = 2
-    OTHER = 3
-
-    @staticmethod
-    def mac_to_vendor(mac: str) -> int:
-        if mac.startswith('e0:62:90'):
-            return VendorType.JOVISION
-        if mac.startswith('e4:24:6c') or mac.startswith('3c:e3:6b'):
-            return VendorType.DAHUA
-        return VendorType.OTHER
+from ... import rosys
+from ...geometry import Rectangle
+from ..camera import Camera, ConfigurableCameraMixin, TransformableCameraMixin
+from ..image import Image, ImageSize
+from ..image_processing import process_jpeg_image
+from ..image_rotation import ImageRotation
+from .rtsp_device import RtspDevice
+from .vendors import VendorType, mac_to_vendor
 
 
 class PeekableBytesIO(io.BytesIO):
@@ -78,126 +59,8 @@ async def find_ip_from_mac(mac: str) -> Optional[str]:
     return None
 
 
-def determine_rtsp_url(mac: str, ip: str, jovision_profile: int = 0) -> Optional[str]:
-    vendor_mac = mac[:8].lower()
-    if vendor_mac == 'e0:62:90':  # Jovision IP Cameras
-        return f'rtsp://admin:admin@{ip}/profile{jovision_profile}'
-    if vendor_mac in ['e4:24:6c', '3c:e3:6b']:  # Dahua IP Cameras
-        return f'rtsp://admin:Adminadmin@{ip}/cam/realmonitor?channel=1&subtype=0'
-    return None
-
-
-class RtspCameraGstreamerDevice:
-    _image_buffer: Optional[bytes] = None
-    capture_task: Optional[asyncio.Task] = None
-    capture_process: Optional[subprocess.Popen] = None
-    rotation: Optional[ImageRotation] = None
-    crop: Optional[Rectangle] = None
-    _authorized: bool = True
-    resolution: Optional[ImageSize] = None
-    settings_interface = Optional[JovisionInterface]
-
-    def __init__(self, mac: str, ip: str, jovision_profile: int) -> None:
-        self.mac = mac
-
-        vendor_type = VendorType.mac_to_vendor(mac)
-        if vendor_type == VendorType.JOVISION:
-            self.settings_interface = JovisionInterface(ip)
-            self.fps = self.settings_interface.get_fps(stream_id=jovision_profile)
-        else:
-            logging.info(f'no settings interface for vendor type {vendor_type}')
-            logging.info(f'using default fps of 10')
-            self.fps = 10
-
-        self.url = determine_rtsp_url(mac, self.ip, jovision_profile)
-        if self.url is None:
-            raise Exception(f'could not determine RTSP URL for {mac}')
-        print(f'connecting to {self.url}', flush=True)
-        logging.info(f'Starting VideoStream for {self.url}')
-        self.capture_task = rosys.background_tasks.create(
-            self.start_gsreamer_task(self.url), name=f'capture {self.mac}')
-
-    @property
-    def authorized(self) -> bool:
-        return self._authorized
-
-    def capture(self) -> Optional[bytes]:
-        image = self._image_buffer
-        self._image_buffer = None
-        return image
-
-    def shutdown(self) -> None:
-        if self.capture_process is not None:
-            self.capture_process.terminate()
-            self.capture_process = None
-
-    async def start_gsreamer_task(self, url: str) -> None:
-        async def stream(url: str) -> AsyncGenerator[bytes, None]:
-            if 'subtype=0' in url:
-                url = url.replace('subtype=0', 'subtype=1')
-
-            # to try: replace avdec_h264 with nvh264dec ! nvvidconv (!videoconvert)
-            command = f'gst-launch-1.0 rtspsrc location="{url}" latency=0 protocols=tcp ! rtph264depay ! avdec_h264 ! videoconvert ! videorate ! "video/x-raw,framerate={self.fps}/1" ! jpegenc ! fdsink'
-            process = await asyncio.create_subprocess_exec(*shlex.split(command), stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-            assert process.stdout is not None
-            assert process.stderr is not None
-            self.capture_process = process
-
-            buffer = BytesIO()
-            pos = 0
-            header = None
-
-            while process.returncode is None:
-                assert process.stdout is not None
-                new = await process.stdout.read(4096)
-                if not new:
-                    break
-                buffer.write(new)
-
-                img_range = None
-                while True:
-                    if header is None:
-                        h = buffer.getvalue().find(b'\xff\xd8', pos)
-                        if h == -1:
-                            pos = buffer.tell() - 1
-                            break
-                        pos = h + 2
-                        header = h
-                    else:
-                        f = buffer.getvalue().find(b'\xff\xd9', pos)
-                        if f == -1:
-                            pos = buffer.tell() - 1
-                            break
-                        img_range = (header, f + 2)
-                        pos = f + 2
-                        header = None
-
-                if img_range:
-                    yield buffer.getvalue()[img_range[0]:img_range[1]]
-
-                    rest = buffer.getvalue()[img_range[1]:]
-                    buffer.seek(0)
-                    buffer.truncate()
-                    buffer.write(rest)
-
-                    pos = 0
-                    if header is not None:
-                        header -= img_range[1]
-
-            assert process.stderr is not None
-            error = await process.stderr.read()
-            logging.info(f'process {process.pid} exited with {process.returncode} and error {error.decode()}')
-            if 'Unauthorized' in error.decode():
-                self._authorized = False
-
-        async for image in stream(url):
-            self._image_buffer = image
-
-        self.capture_task = None
-
-
 class RtspCamera(ConfigurableCameraMixin, TransformableCameraMixin, Camera):
-    device: Optional[RtspCameraGstreamerDevice]
+    device: Optional[RtspDevice]
     jovision_profile: int
 
     detect: bool
@@ -257,7 +120,7 @@ class RtspCamera(ConfigurableCameraMixin, TransformableCameraMixin, Camera):
         if ip is None:
             raise Exception(f'could not find IP address for {self.id}')
 
-        device = RtspCameraGstreamerDevice(mac=self.id, ip=ip, jovision_profile=self.jovision_profile)
+        device = RtspDevice(mac=self.id, ip=ip, jovision_profile=self.jovision_profile)
         if device is None:
             logging.warning(f'could not create device for {self.id}')
             return
@@ -318,7 +181,7 @@ class RtspCamera(ConfigurableCameraMixin, TransformableCameraMixin, Camera):
 
     @staticmethod
     def known_vendor(mac: str) -> bool:
-        return mac[:8] in ['e0:62:90', 'e4:24:6c', '3c:e3:6b']
+        return mac_to_vendor(mac) != VendorType.OTHER
 
     async def _apply_parameters(self, new_values: dict[str, Any]) -> None:
         await super()._apply_parameters(new_values)
